# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "material",
        "contact": {
            "name": "Material",
            "email": "material@mail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/user/me": {
            "post": {
                "tags": [
                    "Api.User"
                ],
                "summary": "我的个人信息",
                "description": "description",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\MeController::index",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginedApiBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "$ref": "#/components/schemas/code"
                                        },
                                        "msg": {
                                            "$ref": "#/components/schemas/msg"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/me"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/me/update": {
            "post": {
                "tags": [
                    "Api.User"
                ],
                "summary": "修改我的个人信息",
                "description": "description",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\MeController::update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "personal_signature": {
                                        "$ref": "#/components/schemas/personal_signature"
                                    }
                                },
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/LoginedApiBody"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "$ref": "#/components/schemas/code"
                                        },
                                        "msg": {
                                            "$ref": "#/components/schemas/msg"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/me"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/refresh-token": {
            "post": {
                "tags": [
                    "Api.User"
                ],
                "summary": "刷新 token（step 3）",
                "description": "当 access_token 过期但 refresh_token 仍在有效期内时，需要请求此接口刷新 token 。如果 refresh_token 也已过期，此接口将响应 code 为 -1002。此时需要用户重新授权",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\RefreshTokenController::index",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "uid": {
                                        "$ref": "#/components/schemas/uid"
                                    },
                                    "sub_uid": {
                                        "$ref": "#/components/schemas/sub_uid"
                                    },
                                    "refresh_token": {
                                        "$ref": "#/components/schemas/refresh_token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "$ref": "#/components/schemas/code"
                                        },
                                        "msg": {
                                            "$ref": "#/components/schemas/msg"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/token_body"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/user": {
            "post": {
                "tags": [
                    "Api.User"
                ],
                "summary": "查看作者的个人信息",
                "description": "description",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\User\\UserController::index",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginedApiBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "$ref": "#/components/schemas/code"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "code": {
                "description": "状态码，0 表示成功，其他表示失败，此时需要将 msg 中的信息通过弹窗展示给用户",
                "type": "integer",
                "example": "0"
            },
            "msg": {
                "description": "提示信息",
                "type": "string",
                "example": "成功"
            },
...
            "sub_uid": {
                "description": "用户子账号id",
                "type": "integer",
                "format": "int32",
                "example": 1
            },
...
            "access_token": {
                "description": "用于验证权限的token,当 token 过期时，请求接口将会响应 code 为 -1001。此时可以请求刷新 token 接口。重新获取 token",
                "type": "string",
                "example": "78657360364647286045dbeb15ba6"
            },
            "refresh_token": {
                "description": "刷新access_token,用于获取新的access_token和refresh_token，并且刷新过期时间",
                "type": "string",
                "example": "78657360364647286045dbeb15f8f"
            },
            "expires_in": {
                "description": "access_token剩余有效时间,单位(秒)",
                "type": "integer",
                "example": "86400"
            },
            "refresh_token_expires_in": {
                "description": "refresh_token剩余有效时间,单位(秒)",
                "type": "integer",
                "example": "2592000"
            },
            "token_body": {
                "properties": {
                    "access_token": {
                        "$ref": "#/components/schemas/access_token"
                    },
                    "refresh_token": {
                        "$ref": "#/components/schemas/refresh_token"
                    },
                    "expires_in": {
                        "$ref": "#/components/schemas/expires_in"
                    },
                    "refresh_token_expires_in": {
                        "$ref": "#/components/schemas/refresh_token_expires_in"
                    }
                },
                "type": "object"
            },
...
            "LoginedApiBody": {
                "description": "已经登陆的 API 接口请求参数",
                "properties": {
                    "access_token": {
                        "$ref": "#/components/schemas/access_token"
                    },
                    "uid": {
                        "$ref": "#/components/schemas/uid"
                    },
                    "sub_uid": {
                        "$ref": "#/components/schemas/sub_uid"
                    }
                },
                "type": "object"
            },
...
            "uid": {
                "description": "用户ID",
                "type": "integer",
                "format": "int32",
                "example": 1
            },
            "avatar": {
                "description": "头像url",
                "type": "string",
                "example": "https://thirdwx.qlogo.cn/mmopen/vi_32/NIAuuyMibmEQ/132"
            },
            "balance": {
                "description": "余额",
                "type": "string",
                "example": "12.23"
            },
            "personal_signature": {
                "description": "用户个性签名",
                "type": "string",
                "example": "我的个性签名"
            },
            "nickname": {
                "description": "用户昵称",
                "type": "string",
                "example": "昵称"
            },
            "user_body": {
                "properties": {
                    "nickname": {
                        "$ref": "#/components/schemas/nickname"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/avatar"
                    },
                    "personal_signature": {
                        "$ref": "#/components/schemas/personal_signature"
                    },
                    "qrcode_url": {
                        "$ref": "#/components/schemas/qrcode_url"
                    }
                },
                "type": "object"
            },
            "user": {
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/uid"
                    }
                },
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/user_body"
                    }
                ]
            },
            "me": {
                "properties": {
                    "total_income": {
                        "$ref": "#/components/schemas/total_income"
                    },
                    "today_income": {
                        "$ref": "#/components/schemas/today_income"
                    },
                    "balance": {
                        "$ref": "#/components/schemas/balance"
                    }
                },
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/user"
                    }
                ]
            },
            "total_income": {
                "description": "出售物料获得的累计收益",
                "type": "string",
                "example": "30.99"
            },
            "today_income": {
                "description": "出售物料获得的今日收益",
                "type": "string",
                "example": "10.99"
            },
            "qrcode_url": {
                "description": "个人主页二维码 url",
                "type": "string",
                "example": "http://api.material.local:80/storage/image/qrcode/2021/06/25/qrcode_1_60d5d40869d9b.jpg"
            },
...
        },
...
    },
...
    "security": [
        []
    ]
}